requirements/base.txt
# Alap csomagok
sentry-sdk
pytz
python-slugify
Pillow
whitenoise
Jinja2
python-dateutil
requests
unidecode

# Django és kiegészítők
django==5.1
django-environ
django-crispy-forms
crispy-bootstrap5
django-bootstrap-datepicker-plus
django-select2-admin-filters
django-select2
django-email-log
django-modeltranslation
django-cleanup
django-filter

# Celery és Redis
celery[redis]
django-celery-results
django-celery-beat

# Fájlkezelés és PDF generálás
python-magic
pdfkit

# Fizetés (Stripe)
stripe

# Google API integrációk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django-cors-headers
